% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_daily_counts.R
\name{get_daily_counts}
\alias{get_daily_counts}
\title{get_daily_counts
Obtiene cuentas de retiros y arribos por estación, además de sumas en ventanas de los últimos 60 minutos de cada una
Los últimos 60 minutos representan que tan concurrida ha sido esta estación recientemente.
La función está vectorizada y buscará todas las combinaciones month y year requeridas
Solo buscar meses que se requieran ya que puede llegar a ser muy intensiva en RAM.}
\usage{
get_daily_counts(month = NULL, year = NULL)
}
\arguments{
\item{month}{mes o vector de meses requeridos}

\item{year}{mes o vector de meses requeridos}
}
\value{

}
\description{
get_daily_counts
Obtiene cuentas de retiros y arribos por estación, además de sumas en ventanas de los últimos 60 minutos de cada una
Los últimos 60 minutos representan que tan concurrida ha sido esta estación recientemente.
La función está vectorizada y buscará todas las combinaciones month y year requeridas
Solo buscar meses que se requieran ya que puede llegar a ser muy intensiva en RAM.
}
\examples{
get_daily_counts(9,2019)
}
