% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_helpers.R
\name{get_daily_counts_sf}
\alias{get_daily_counts_sf}
\title{get_daily_counts_sf}
\usage{
get_daily_counts_sf(month = NULL, year = NULL)
}
\arguments{
\item{month}{mes o vector de meses requeridos}

\item{year}{mes o vector de meses requeridos}
}
\value{
df dataframe con todos los viajes
}
\description{
Igual a get_daily_counts pero regresa convertido a sf
}
