% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-get_month.R
\name{get_trips}
\alias{get_trips}
\title{get_trips}
\usage{
get_trips(month, year)
}
\arguments{
\item{month}{mes o vector de meses requeridos}

\item{year}{mes o vector de meses requeridos}
}
\value{
df dataframe con todos los viajes
}
\description{
Descarga los viajes del sistema Ecobici de la CDMX para los meses deseados.
La función buscará todas las combinaciones month y year requeridas
Solo buscar meses que se requieran ya que puede llegar a ser muy intensiva en RAM.
}
