% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-update_api.R
\name{update_ecobici_estaciones}
\alias{update_ecobici_estaciones}
\title{update_ecobici_estaciones
Actualiza el índice de estaciones que se debe de consultar a la API de Ecobici}
\usage{
update_ecobici_estaciones()
}
\arguments{
\item{client_secret}{}
}
\value{

}
\description{
update_ecobici_estaciones
Actualiza el índice de estaciones que se debe de consultar a la API de Ecobici
}
