% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-update_api.R
\name{update_ecobici_status}
\alias{update_ecobici_status}
\title{update_ecobici_status
Actualiza dos dataframes que se deben de consultar a la API de Ecobici,
el índice de estaciones}
\usage{
update_ecobici_status()
}
\arguments{
\item{client_secret}{}
}
\value{

}
\description{
update_ecobici_status
Actualiza dos dataframes que se deben de consultar a la API de Ecobici,
el índice de estaciones
}
